requirements.txt
openai>=0.27.0
google-cloud-bigquery>=3.0.0
faiss-cpu>=1.7.3
numpy>=1.23.0
pandas>=1.5.0
streamlit>=1.19.0
pytest>=7.0.0
redis>=4.3.0           # optional, if using Redis vector store
redis-om>=0.1.0        # optional, if using Redis OM
pipenv>=2023.7.7       # or virtualenv/venv of your choice
sentence-transformers>=2.2.0  # for open-source embeddings
pyyaml>=6.0
fastapi
uvicorn
structlog
langchain>=0.1.14
langchain-community>=0.0.24
google-cloud-aiplatform